# Conflicts:
#	EmberApp/app/templates/dashboard/home.hbs

+++ src/UserBundle/Business/Manager/Agent/JsonApiUpdateAgentManagerTrait.php
@@ -62,16 +62,16 @@ trait JsonApiUpdateAgentManagerTrait
                 $this->eventDispatcher->dispatch(AgentEvents::ON_AGENT_GROUP_CHANGE, $event);
             }
 
-//            try {
-//                $syncResult = $this->syncWithTCRPortal($agent, 'edit');
-//                if (is_object($syncResult) && $syncResult->code == 200) {
+            try {
+                $syncResult = $this->syncWithTCRPortal($agent, 'edit');
+                if (is_object($syncResult) && $syncResult->code == 200) {
                     $this->flushDb();
-//                } else {
-//                    return new ArrayCollection(AgentApiResponse::AGENT_SYNC_ERROR_RESPONSE);
-//                }
-//            } catch (\Exception $exception) {
-//                return new ArrayCollection(AgentApiResponse::AGENT_SYNC_ERROR_RESPONSE);
-//            }
+                } else {
+                    return new ArrayCollection(AgentApiResponse::AGENT_SYNC_ERROR_RESPONSE);
+                }
+            } catch (\Exception $exception) {
+                return new ArrayCollection(AgentApiResponse::AGENT_SYNC_ERROR_RESPONSE);
+            }
         }
 
         return $this->createJsonAPiUpdateResponse($agentOrException);
@@ -86,7 +86,7 @@ trait JsonApiUpdateAgentManagerTrait
         $this->setAndValidateAddress($agent, $dbAgent);
         $this->setAndValidateGroup($agent, $dbAgent);
         $this->setAndValidateImage($agent, $dbAgent);
-
+        $this->setAndValidateNotifications($agent, $dbAgent);
 
         return $agent;
     }
@@ -153,6 +153,11 @@ trait JsonApiUpdateAgentManagerTrait
         $agent->getGroup() ? ($dbGroup = $this->groupManager->getReference($agent->getGroup()->getId()))&& $dbAgent->setGroup($dbGroup):false;
 
 
+    }
+
+    private function setAndValidateNotifications (Agent $agent, Agent $dbAgent)
+    {
+        $dbAgent->setNotifications($agent->getNotifications());
     }
